Move Plugins to its own package
