+++ src/main/kotlin/guru/stefma/androidartifacts/plugin/AndroidArtifactsPlugin.kt
@@ -1,6 +1,7 @@
-package guru.stefma.androidartifacts
+package guru.stefma.androidartifacts.plugin
 
 import com.android.build.gradle.api.LibraryVariant
+import guru.stefma.androidartifacts.*
 import org.gradle.api.Plugin
 import org.gradle.api.Project
 import org.gradle.api.publish.PublicationContainer
