+++ src/main/kotlin/guru/stefma/androidartifacts/plugin/JavaArtifactsPlugin.kt
@@ -1,5 +1,6 @@
-package guru.stefma.androidartifacts
+package guru.stefma.androidartifacts.plugin
 
+import guru.stefma.androidartifacts.*
 import org.gradle.api.Plugin
 import org.gradle.api.Project
 import org.gradle.api.publish.maven.MavenPublication
